+++ src/Api/AbstractApi.php
@@ -14,6 +14,7 @@ use Http\Message\MultipartStream\MultipartStreamBuilder;
 use Psr\Http\Message\StreamFactoryInterface;
 use Psr\Http\Message\StreamInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use ValueError;
 
 /**
  * @author Joseph Bielawski <stloyd@gmail.com>
@@ -30,28 +31,57 @@ abstract class AbstractApi implements ApiInterface
     private const URI_PREFIX = '/api/v4/';
 
     /**
-     * The HTTP methods client.
+     * The client instance.
      *
-     * @var HttpMethodsClientInterface
+     * @var Client
      */
-    private $httpClient;
+    private $client;
 
     /**
-     * The HTTP stream factory.
+     * The per page parameter.
      *
-     * @var StreamFactoryInterface
+     * @var int|null
      */
-    private $streamFactory;
+    private $perPage;
 
     /**
-     * @param Client $client
+     * Create a new API instance.
+     *
+     * @param Client   $client
+     * @param int|null $perPage
      *
      * @return void
      */
-    public function __construct(Client $client)
+    public function __construct(Client $client, ?int $perPage)
     {
-        $this->httpClient = $client->getHttpClient();
-        $this->streamFactory = $client->getStreamFactory();
+        if (null !== $perPage && ($perPage < 1 || $perPage > 100)) {
+            throw new ValueError(sprintf('%s::__construct(): Argument #2 ($perPage) must be between 1 and 100, or null', self::class));
+        }
+
+        $this->client = $client;
+        $this->perPage = $perPage;
+    }
+
+    /**
+     * Create a new instance with the given page parameter.
+     *
+     * This must be an integer between 1 and 100.
+     *
+     * @param int|null $perPage
+     *
+     * @return static
+     */
+    public function perPage(?int $perPage)
+    {
+        if (null !== $perPage && ($perPage < 1 || $perPage > 100)) {
+            throw new ValueError(sprintf('%s::perPage(): Argument #1 ($perPage) must be between 1 and 100, or null', self::class));
+        }
+
+        $copy = clone $this;
+
+        $copy->perPage = $perPage;
+
+        return $copy;
     }
 
     /**
@@ -67,7 +97,11 @@ abstract class AbstractApi implements ApiInterface
      */
     protected function getAsResponse($uri, array $params = [], array $headers = [])
     {
-        return $this->httpClient->get(self::prepareUri($uri, $params), $headers);
+        if (null !== $this->perPage && !isset($params['per_page'])) {
+            $params['per_page'] = $this->perPage;
+        }
+
+        return $this->client->getHttpClient()->get(self::prepareUri($uri, $params), $headers);
     }
 
     /**
@@ -106,7 +140,7 @@ abstract class AbstractApi implements ApiInterface
             }
         }
 
-        $response = $this->httpClient->post(self::prepareUri($uri), $headers, $body);
+        $response = $this->client->getHttpClient()->post(self::prepareUri($uri), $headers, $body);
 
         return ResponseMediator::getContent($response);
     }
@@ -133,7 +167,7 @@ abstract class AbstractApi implements ApiInterface
             }
         }
 
-        $response = $this->httpClient->put(self::prepareUri($uri), $headers, $body ?? '');
+        $response = $this->client->getHttpClient()->put(self::prepareUri($uri), $headers, $body ?? '');
 
         return ResponseMediator::getContent($response);
     }
@@ -153,7 +187,7 @@ abstract class AbstractApi implements ApiInterface
             $headers = self::addJsonContentType($headers);
         }
 
-        $response = $this->httpClient->delete(self::prepareUri($uri), $headers, $body ?? '');
+        $response = $this->client->getHttpClient()->delete(self::prepareUri($uri), $headers, $body ?? '');
 
         return ResponseMediator::getContent($response);
     }
@@ -241,7 +275,7 @@ abstract class AbstractApi implements ApiInterface
      */
     private function createMultipartStreamBuilder(array $params = [], array $files = [])
     {
-        $builder = new MultipartStreamBuilder($this->streamFactory);
+        $builder = new MultipartStreamBuilder($this->client->getStreamFactory());
 
         foreach ($params as $name => $value) {
             $builder->addResource($name, $value);