+++ src/Api/ApiInterface.php
@@ -6,10 +6,16 @@ namespace Gitlab\Api;
 
 use Gitlab\Client;
 
-/**
- * Api interface.
- */
 interface ApiInterface
 {
-    public function __construct(Client $client);
+    /**
+     * Create a new instance with the given per page parameter.
+     *
+     * This must be an integer between 1 and 100.
+     *
+     * @param int|null $perPage
+     *
+     * @return static
+     */
+    public function perPage(?int $perPage);
 }