+++ src/ResultPager.php
@@ -5,8 +5,10 @@ declare(strict_types=1);
 namespace Gitlab;
 
 use Gitlab\Api\ApiInterface;
+use Gitlab\Client;
 use Gitlab\Exception\RuntimeException;
 use Gitlab\HttpClient\Message\ResponseMediator;
+use ValueError;
 
 /**
  * This is the result pager class.
@@ -18,11 +20,25 @@ use Gitlab\HttpClient\Message\ResponseMediator;
 final class ResultPager implements ResultPagerInterface
 {
     /**
+     * The default number of entries to request per page.
+     *
+     * @var int
+     */
+    private const PER_PAGE = 50;
+
+    /**
      * The client to use for pagination.
      *
-     * @var \Gitlab\Client client
+     * @var Client
      */
     private $client;
+
+    /**
+     * The number of entries to request per page.
+     *
+     * @var int
+     */
+    private $perPage;
 
     /**
      * The pagination result from the API.
@@ -34,13 +50,19 @@ final class ResultPager implements ResultPagerInterface
     /**
      * Create a new result pager instance.
      *
-     * @param \Gitlab\Client $client
+     * @param Client   $client
+     * @param int|null $perPage
      *
      * @return void
      */
-    public function __construct(Client $client)
+    public function __construct(Client $client, int $perPage = null)
     {
+        if (null !== $perPage && ($perPage < 1 || $perPage > 200)) {
+            throw new ValueError(sprintf('%s::__construct(): Argument #2 ($perPage) must be between 1 and 100, or null', self::class));
+        }
+
         $this->client = $client;
+        $this->perPage = $perPage ?? self::PER_PAGE;
         $this->pagination = [];
     }
 
@@ -57,10 +79,10 @@ final class ResultPager implements ResultPagerInterface
      */
     public function fetch(ApiInterface $api, string $method, array $parameters = [])
     {
-        $result = $api->$method(...$parameters);
+        $result = $api->perPage($this->perPage)->$method(...$parameters);
 
         if (!is_array($result)) {
-            throw new RuntimeException('Pagination of endpoints that produce blobs is not supported.');
+            throw new RuntimeException('Pagination of this endpoint is not supported.');
         }
 
         $this->postFetch();
@@ -81,13 +103,33 @@ final class ResultPager implements ResultPagerInterface
      */
     public function fetchAll(ApiInterface $api, string $method, array $parameters = [])
     {
-        $result = $this->fetch($api, $method, $parameters);
+        return iterator_to_array($this->fetchAllLazy($api, $method, $parameters));
+    }
 
-        while ($this->hasNext()) {
-            $result = array_merge($result, $this->fetchNext());
+    /**
+     * Lazily fetch all results from an api call.
+     *
+     * @param ApiInterface $api
+     * @param string       $method
+     * @param array        $parameters
+     *
+     * @throws \Http\Client\Exception
+     *
+     * @return \Generator
+     */
+    public function fetchAllLazy(ApiInterface $api, string $method, array $parameters = [])
+    {
+        /** @var mixed $value */
+        foreach ($this->fetch($api, $method, $parameters) as $value) {
+            yield $value;
         }
 
-        return $result;
+        while ($this->hasNext()) {
+            /** @var mixed $value */
+            foreach ($this->fetchNext() as $value) {
+                yield $value;
+            }
+        }
     }
 
     /**
@@ -194,7 +236,7 @@ final class ResultPager implements ResultPagerInterface
         $content = ResponseMediator::getContent($result);
 
         if (!is_array($content)) {
-            throw new RuntimeException('Pagination of endpoints that produce blobs is not supported.');
+            throw new RuntimeException('Pagination of this endpoint is not supported.');
         }
 
         $this->postFetch();