+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGE LOG
 * Switched to PSR-17 and PSR-18
 * Encode URIs according to RFC 3986
 * Send request bodies as JSON to GitLab
+* Redesigned pagination
 
 
 ## 9.18.0 (11/07/2020)